ParallelPopen mostly works

+++ psshlib/basethread.py
@@ -24,33 +24,59 @@ class ParallelPopen(object):
         self.done = []
 
     def add_task(self, task):
-        if len(self.running) < self.limit:
-            self.running.append(task)
-            task.start()
-        else:
-            self.tasks.append(task)
+        self.tasks.append(task)
 
-    def wait(self):
+    def run(self):
+        writer = Writer()
+        writer.start()
+
         try:
-            start_tasks()
-            while True:
-                self.iomap.poll()
+            self.start_tasks(writer)
+            timeout = None
+            while self.running or self.tasks:
+                self.iomap.poll(timeout)
                 self.check_tasks()
+                timeout = self.check_timeout()
         except KeyboardInterrupt:
             self.stop_all()
+
+        # TODO: finish task cleanup here
+
+        writer.queue.put((Writer.ABORT, None))
+        writer.join()
+
+    def start_tasks(self, writer):
+        while 0 < len(self.tasks) and len(self.running) < self.limit:
+            task = self.tasks.pop(0)
+            self.running.append(task)
+            task.start(self.iomap, writer)
 
     def stop_all(self):
         for task in self.running:
             task.stop()
 
     def check_tasks(self):
+        still_running = []
         for task in self.running:
-            pass
+            if task.running():
+                still_running.append(task)
+            else:
+                self.done.append(task)
+                n = len(self.done)
+                task.report(n)
+        self.running = still_running
 
-        while 0 < len(self.tasks) and len(self.running) < self.limit:
-            task = self.tasks.pop(0)
-            self.running.append(task)
-            task.start()
+    def check_timeout(self):
+        """Kills timed-out processes and returns the lowest time left."""
+        min_timeleft = None
+        for task in self.running:
+            timeleft = task.timeleft()
+            if timeleft <= 0:
+                task.stop()
+                continue
+            if min_timeleft is None or timeleft < min_timeleft:
+                min_timeleft = timeleft
+        return max(0, min_timeleft)
 
 
 class Task(object):
@@ -58,18 +84,14 @@ class Task(object):
         self.host = host
         self.port = port
         self.cmd = cmd
-        self.stdin = stdin
-        self.outputbuffer = ""
-        self.errorbuffer = ""
-        self.exc_info = []
-        self.exc_str = []
 
         self.proc = None
-        self.iomap = None
         self.writer = None
         self.timestamp = None
 
         # Set options.
+        self.timeout = opts.timeout
+        self.verbose = opts.verbose
         self.outdir = opts.outdir
         self.errdir = opts.errdir
         try:
@@ -81,6 +103,17 @@ class Task(object):
         except AttributeError:
             self.inline = False
 
+        self.stdin = None
+        self.stdout = None
+        self.stderr = None
+        self.inputbuffer = stdin
+        self.outputbuffer = ''
+        self.outfile = None
+        self.errorbuffer = ''
+        self.errfile = None
+        self.exc_info = []
+        self.exc_str = []
+
     def start(self, iomap, writer):
         self.writer = writer
 
@@ -89,57 +122,66 @@ class Task(object):
         else:
             stdin = None
 
-        stdout = stderr = None
         if self.outdir:
             pathname = "%s/%s" % (self.outdir, self.host)
             self.outfile = open(pathname, "w")
-            stdout = PIPE
         if self.errdir:
             pathname = "%s/%s" % (self.errdir, self.host)
             self.errfile = open(pathname, "w")
-            stderr = PIPE
-        if self.inline or self.print_out:
-            stdout = stderr = PIPE
 
         # Create the subprocess.
-        self.proc = Popen(self.cmd, stderr=stderr, stdin=stdin, stdout=stdout,
-                close_fds=True, preexec_fn=os.setsid)
+        self.proc = Popen([self.cmd], stdin=PIPE, stdout=PIPE, stderr=PIPE,
+                close_fds=True, preexec_fn=os.setsid, shell=True)
         self.timestamp = time.time()
         if stdin:
-            self.fileno_stdin = self.proc.stdin.fileno()
-            iomap.register(self.fileno_stdin, self.handle_stdin, write=True)
-        if stdout:
-            self.fileno_stdout = self.proc.stdout.fileno()
-            iomap.register(self.fileno_stdout, self.handle_stdout, read=True)
-        if stderr:
-            self.fileno_stderr = self.proc.stderr.fileno()
-            iomap.register(self.fileno_stderr, self.handle_stderr, read=True)
+            self.stdin = self.proc.stdin
+            iomap.register(self.stdin.fileno(), self.handle_stdin, write=True)
+        else:
+            self.proc.stdin.close()
+        self.stdout = self.proc.stdout
+        iomap.register(self.stdout.fileno(), self.handle_stdout, read=True)
+        self.stderr = self.proc.stderr
+        iomap.register(self.stderr.fileno(), self.handle_stderr, read=True)
 
     def stop(self):
-        os.kill(-self.proc.pid, signal.SIGKILL)
-        # TODO: save a message that notes that this was interrupted!
+        if self.proc:
+            os.kill(-self.proc.pid, signal.SIGKILL)
+            # TODO: save a message that notes that this was interrupted!
 
     def timeleft(self):
         if self.timeout and self.timeout > 0:
             return self.timeout - (time.time() - self.timestamp)
 
-    def handle_stdin(self, fd, event):
+    def running(self):
+        if self.stdin or self.stdout or self.stderr:
+            return True
+        if self.proc:
+            self.returncode = self.proc.poll()
+            if self.returncode is None:
+                return True
+            else:
+                print 'task done running'
+                self.proc = None
+                return False
+
+    def handle_stdin(self, fd, event, iomap):
         try:
             bytes_written = os.write(fd, self.stdin)
-        except:
-            self.close_stdin()
+        except (OSError, IOError), e:
+            self.close_stdin(iomap)
             self.log_exception(e)
         self.stdin = self.stdin[bytes_written:]
 
-    def close_stdin(self):
-        if self.fileno_stdin:
-            self.iomap.unregister(self.fileno_stdin)
-            os.close(self.fileno_stdin)
-            self.fileno_stdin = None
+    def close_stdin(self, iomap):
+        if self.stdin:
+            iomap.unregister(self.stdin.fileno())
+            self.stdin.close()
+            self.stdin = None
 
-    def handle_stdout(self, fd, event):
+    def handle_stdout(self, fd, event, iomap):
         try:
             buf = os.read(fd, BUFFER_SIZE)
+            #print 'got from stdout:', buf
             if buf:
                 if self.inline:
                     self.outputbuffer += buf
@@ -148,21 +190,23 @@ class Task(object):
                 if self.print_out:
                     print '%s: %s' % (self.host, buf),
             else:
-                self.close_stdout()
+                #print 'EOF, outputbuffer:', self.outputbuffer
+                self.close_stdout(iomap)
         except (OSError, IOError), e:
-            self.close_stdout()
+            self.close_stdout(iomap)
             self.log_exception(e)
 
-    def close_stdout(self):
-        if self.fileno_stdout:
-            self.iomap.unregister(self.fileno_stdout)
-            os.close(self.fileno_stdout)
-            self.fileno_stdout = None
+    def close_stdout(self, iomap):
+        #print 'closing stdout'
+        if self.stdout:
+            iomap.unregister(self.stdout.fileno())
+            self.stdout.close()
+            self.stdout = None
         if self.outfile:
             self.writer.write(self.outfile, Writer.EOF)
             self.outfile = None
 
-    def handle_stderr(self, fd, event):
+    def handle_stderr(self, fd, event, iomap):
         try:
             buf = os.read(fd, BUFFER_SIZE)
             if buf:
@@ -171,25 +215,26 @@ class Task(object):
                 if self.errfile:
                     self.writer.write(self.errfile, buf)
             else:
-                self.close_stderr()
-        except Exception, e:
-            self.close_stderr()
+                self.close_stderr(iomap)
+        except (OSError, IOError), e:
+            self.close_stderr(iomap)
             self.log_exception(e)
 
-    def close_stderr(self):
-        if self.fileno_stderr:
-            self.iomap.unregister(self.fileno_stderr)
-            os.close(self.fileno_stderr)
-            self.fileno_stderr = None
+    def close_stderr(self, iomap):
+        if self.stderr:
+            iomap.unregister(self.stderr.fileno())
+            self.stderr.close()
+            self.stderr = None
         if self.errfile:
             self.writer.write(self.errfile, Writer.EOF)
             self.errfile = None
 
     def log_exception(self, e):
+        print 'got an exception'
         self.exc_info.append(sys.exc_info())
         self.exc_str.append(str(e))
 
-    def log_completion(self, n):
+    def report(self, n):
         tstamp = time.asctime().split()[3] # Current time
         if self.verbose:
             exceptions = []
@@ -203,14 +248,14 @@ class Task(object):
             progress = color.c("[%s]" % color.B(n))
             success = color.g("[%s]" % color.B("SUCCESS"))
             failure = color.r("[%s]" % color.B("FAILURE"))
-            stderr = color.r("Standard error:"))
+            stderr = color.r("Standard error:")
             exc = color.r(color.B(exc))
         else:
             progress = "[%s]" % n
             success = "[SUCCESS]"
             failure = "[FAILURE]"
             stderr = "Standard error:"
-        if exceptions
+        if self.exc_info:
             print progress, tstamp, failure, self.host, self.port, exc
         else:
             print progress, tstamp, success, self.host, self.port
@@ -223,9 +268,11 @@ class Task(object):
 
 class Writer(threading.Thread):
     EOF = object()
+    ABORT = object()
 
     def __init__(self):
         threading.Thread.__init__(self)
+        self.setDaemon(True)
         self.queue = Queue.Queue()
 
     def write(self, fd, data):
@@ -235,7 +282,9 @@ class Writer(threading.Thread):
     def run(self):
         while True:
             file, data = self.queue.get()
-            if file == self.EOF:
+            if file == self.ABORT:
+                return
+            if data == self.EOF:
                 file.close()
             else:
                 print >>file, data,
@@ -271,5 +320,5 @@ class IOMap(object):
         """Performs a poll and dispatches the resulting events."""
         for fd, event in self.poller.poll(timeout):
             handler = self.map[fd]
-            handler(fd, event)
+            handler(fd, event, self)
